Checkpoint before huge update Core Extension:     - Added "timeFormated" which takes miliseconds and returns a         nice formated and readable string (eg. "1 day, 23 seconds") Mafia Module:     - Added "replace" command to swap out an inactive player       wihoutleaving 'dead' votes     - Added "showProdTimers" command - does exactly that     - Added "timeleft" command which showes the time left till phase end     - Added "updateProdtimers" event which does that on player messages       in a game channel     - Added "check_prods" task which sends out messages to mods if a       prod timer is exceeded     - Modified "startPhase" command to fix some bugs     - Modified "mafia_menu" in general to work again     - Updated readme     - Moved "getGamesFromContext" into "gameHelper"
Current issue:
With how the events and task work currently it can happen that the db
runs into race conditions which is especially (well actually currently
the only place where I noticed it) noticeable in the command "startPhase".
It sometimes starts the next phase but because of the task "updateProdtimers"
it writes the old game back to the db with the phase not jet updated.
So I will rewrite all events and taks to be processed one at a time one
after another.